Removed ifixpng() and fixed expand()
The new set of SVG icons in this fork have removed the need for
ifixpng(). It has been removed to reduce the JS payload.

Fixed and reinstated the expand() function. All kinds of problems here:
failing conditonals, and absolute setting of element sizes that causes
mayhem if the user resizes the browser window. All fixed now and
working in a responsive-design-friendly manner.